% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndset.R
\name{classifyNDSet}
\alias{classifyNDSet}
\title{Classify a set of nondominated points}
\usage{
classifyNDSet(pts, direction = 1)
}
\arguments{
\item{pts}{A set of non-dominated points. It is assumed that \code{ncol(pts)} equals the number of
objectives ($p$).}
}
\value{

}
\description{
The classification is supported (true/false), extreme (true/false), supported non-extreme
(true/false)
}
\note{
It is assumed that \code{pts} are nondominated.
}
\examples{
pts <- genNDSet(3,50)
ini3D(argsPlot3d = list(xlim = c(0,max(pts$x)+2),
  ylim = c(0,max(pts$y)+2),
  zlim = c(0,max(pts$z)+2)))
plotHull3D(pts, addR3 = T, argsPolygon3d = list(alpha = 0.5))
pts <- classifyNDSet(pts[,1:3])
plotPoints3D(pts[pts$se,1:3], argsPlot3d = list(col = "red"))
plotPoints3D(pts[!pts$sne,1:3], argsPlot3d = list(col = "black"))
plotPoints3D(pts[!pts$us,1:3], argsPlot3d = list(col = "blue"))
finalize3D()

}
