% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hull.R
\name{convexHull3D}
\alias{convexHull3D}
\title{Find the convex hull of a set of points in 3D.}
\usage{
convexHull3D(pts, classify = FALSE, addR3 = FALSE)
}
\arguments{
\item{pts}{A matrix with a point in each row.}

\item{classify}{If true, a classification the input points are given (a column)}

\item{addR3}{Add the positive cone to the hull}
}
\value{
If \code{classify = FALSE}, a vector/matrix with row indices of the vertices defining
each facet in the hull \code{pts}; Otherwise, a list with \code{hull} equal the same result as
under \code{classify = FALSE} and \code{pts} equal the input points with two additional rows:
\code{pt}, true if a point in the original input; false otherwise. \code{vtx}, TRUE if a
vertex in the hull.
}
\description{
Find the convex hull of a set of points in 3D.
}
\examples{
pts<-matrix(c(1,1,1, 1,1,1), ncol = 3, byrow = TRUE)
dimFace(pts) # a point
convexHull3D(pts)
pts<-matrix(c(0,0,0,1,1,1,2,2,2,3,3,3), ncol = 3, byrow = TRUE)
dimFace(pts) # a line
convexHull3D(pts)
pts<-matrix(c(0,0,0,0,1,1,0,2,2,0,0,2), ncol = 3, byrow = TRUE)
dimFace(pts) # a polygon
convexHull3D(pts)
pts<-matrix(c(1,0,0,1,1,1,1,2,2,3,1,1), ncol = 3, byrow = TRUE)
dimFace(pts) # a polygon
convexHull3D(pts, classify = TRUE) # a polyhedron
pts<-matrix(c(1,1,1,2,2,1,2,1,1,1,1,2), ncol = 3, byrow = TRUE)
dimFace(pts) # a polytope (polyhedron)
convexHull3D(pts)

ini3D(argsPlot3d = list(xlim = c(0,3), ylim = c(0,3), zlim = c(0,3)))
pts<-matrix(c(1,1,1,2,2,1,2,1,1,1,1,2), ncol = 3, byrow = TRUE)
plotPoints3D(pts)
plotHull3D(pts, argsPolygon3d = list(color = "red"))
convexHull3D(pts, classify = TRUE, addR3 = FALSE)
plotHull3D(pts, addR3 = TRUE)
convexHull3D(pts, classify = TRUE, addR3 = TRUE)
finalize3D()
}
