% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{df2String}
\alias{df2String}
\title{Convert each row to a string.}
\usage{
df2String(df, round = 2)
}
\arguments{
\item{df}{Data frame.}

\item{round}{How many digits to round}
}
\value{
A vector of strings.
}
\description{
Convert each row to a string.
}
