% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.checkPts}
\alias{.checkPts}
\title{Check if point input is okay}
\usage{
.checkPts(pts, p = NULL, warn = FALSE)
}
\arguments{
\item{pts}{Point input.}

\item{p}{Desired dimension of points.}

\item{warn}{Output warnings.}
}
\value{
Point input converted to a matrix.
}
\description{
Check if point input is okay
}
