% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.getRanges}
\alias{.getRanges}
\title{Get ranges of the bounding box margins}
\usage{
.getRanges(expand = 1.03, ranges = par3d("bbox"))
}
\arguments{
\item{expand}{Expand margins.}

\item{ranges}{The bounding box.}
}
\value{
A list with ranges.
}
\description{
Get ranges of the bounding box margins
}
