% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{.mToDirection}
\alias{.mToDirection}
\title{Convert min/max to direction (1/-1)}
\usage{
.mToDirection(m, p)
}
\arguments{
\item{m}{Min/max vector.}

\item{p}{Number of objectives.}
}
\value{
A direction vector (min = 1 and max = -1)
}
\description{
Convert min/max to direction (1/-1)
}
\keyword{internal}
