% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{.sizeM}
\alias{.sizeM}
\title{Estimate 1 em in pixels in the resulting png.}
\usage{
.sizeM(...)
}
\arguments{
\item{...}{Arguments parsed on to \code{texToPng}.}
}
\value{
The width and size of the png.
}
\description{
Estimate 1 em in pixels in the resulting png.
}
