% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{finalize3D}
\alias{finalize3D}
\title{Finalize the rgl window.}
\usage{
finalize3D(...)
}
\arguments{
\item{...}{Further arguments passed on the the rgl plotting functions. This must be done as
lists. Currently the following arguments are supported:
\itemize{
\item \code{argsAxes3d}: A list of arguments for \code{\link[rgl:axes3d]{rgl::axes3d}}.
\item \code{argsTitle3d}: A list of arguments for \code{\link[rgl:axes3d]{rgl::title3d}}.
}}
}
\value{
NULL (invisible).
}
\description{
Finalize the rgl window.
}
\examples{
ini3D()
pts<-matrix(c(1,1,1,5,5,5), ncol = 3, byrow = TRUE)
plotPoints3D(pts)
finalize3D()

ini3D()
pts<-matrix(c(1,1,1,5,5,5), ncol = 3, byrow = TRUE)
plotPoints3D(pts)
finalize3D(argsAxes3d = list(edges = "bbox"))
}
