% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getA}
\alias{getA}
\title{Return the A matrix of the LP}
\usage{
getA(LP)
}
\arguments{
\item{LP}{The LP object defined using lpSolveAPI}
}
\description{
Return the A matrix of the LP
}
\author{
Lars Relund \email{lars@relund.dk}
}

