% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getC}
\alias{getC}
\title{Return the coefficient vector of the LP}
\usage{
getC(LP)
}
\arguments{
\item{LP}{The LP object defined using lpSolveAPI}
}
\description{
Return the coefficient vector of the LP
}
\author{
Lars Relund \email{lars@relund.dk}
}

