% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_doc.R
\docType{package}
\name{gMOIP-package}
\alias{gMOIP}
\alias{gMOIP-package}
\title{gMOIP: Tools for 2D and 3D Plots of Single and Multi-Objective Linear/Integer Programming Models}
\description{
Make 2D and 3D plots of linear programming (LP), integer linear programming (ILP), or mixed integer linear programming (MILP) models with up to three objectives. Plots of both the solution and criterion space are possible. For instance the non-dominated (Pareto) set for bi-objective LP/ILP/MILP programming models (see vignettes for an overview). The package also contains an function for checking if a point is inside the convex hull.
}
\seealso{
\code{\link[=plotPolytope]{plotPolytope()}}, \code{\link[=plotCriterion2D]{plotCriterion2D()}} and \code{\link[=plotHull3D]{plotHull3D()}}.
}
\author{
\strong{Maintainer}: Lars Relund Nielsen \email{lars@relund.dk} (\href{https://orcid.org/0000-0002-4802-3071}{ORCID})

}
\keyword{internal}
