% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_doc.R
\docType{package}
\name{gMOIP}
\alias{gMOIP}
\alias{gMOIP-package}
\title{2D plots of linear/integer programming models.}
\description{
Make 2D plots of the polyeder of a LP or IP problem, including integer points and iso profit
curve. Can also make a plot of a bi-objective criterion space.
}
\seealso{
\link{plotPolytope} and \link{plotCriterion2D}.
}
\author{
Lars Relund <lars@relund.dk> [aut, cre].
}
\keyword{package}
