% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gMOIPTheme}
\alias{gMOIPTheme}
\title{The \code{ggplot} theme for the package}
\usage{
gMOIPTheme(...)
}
\arguments{
\item{...}{Further arguments parsed to \code{\link[ggplot2:theme]{ggplot2::theme()}}.}
}
\value{
The theme object.
}
\description{
The \code{ggplot} theme for the package
}
