% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMTeX3D}
\alias{plotMTeX3D}
\title{Plot TeX in the margin}
\usage{
plotMTeX3D(tex, edge, line = 0, at = NULL, pos = NA, ...)
}
\arguments{
\item{tex}{TeX string}

\item{edge}{The position at which to draw the axis or text.}

\item{line}{The ``line'' of the plot margin to draw the label on.}

\item{at}{The value of a coordinate at which to draw the axis.}

\item{pos}{The position at which to draw the axis or text.}

\item{...}{Further arguments passed to \code{\link[=plotTeX3D]{plotTeX3D()}}.}
}
\value{
The object IDs of objects added to the scene.
}
\description{
Plot TeX in the margin
}
