% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ndset.R
\name{genNDSet}
\alias{genNDSet}
\title{Generate a sample of nondominated points.}
\usage{
genNDSet(n = 100, random = FALSE, onSphere = TRUE, keep = F, ...)
}
\arguments{
\item{n}{Number of samples generated (note only a subset of these points are nondominated).}

\item{random}{Random sampling.}

\item{onSphere}{Generate points on a sphere.}

\item{keep}{Keep dominated points also.}

\item{...}{Arguments for generating the points based on method. Currently two lists may be added:
argsRandom Arguments for random sampling.
argsSphere Arguments for sampling on a sphere.}
}
\value{
A data frame with xyz columns and ...0
}
\description{
Generate a sample of nondominated points.
}
\examples{
ini3D()
p <- genNDSet(1000, random = TRUE, keep = TRUE)
head(p)
plotPoints3D(p)
plotPoints3D(p[!p$dom,], argsPlot3d = list(col = "red", size = 10))
finalize3D()

ini3D()
p <- genNDSet(1000, keep = TRUE)
rgl::spheres3d(c(52,52,52), radius=50, color = "grey100", alpha=0.1)
plotPoints3D(p)
plotPoints3D(p[!p$dom,], argsPlot3d = list(col = "red", size = 10))
rgl::planes3d(52,52,52,-8112, alpha = 0.5, col = "red")
finalize3D()

ini3D()
cent <- c(100,100,100)
r <- 75
p <- genNDSet(1000, keep = TRUE, argsSphere = list(center = cent, radius = r))
rgl::spheres3d(cent, radius=r, color = "grey100", alpha=0.1)
plotPoints3D(p)
plotPoints3D(p[!p$dom,], argsPlot3d = list(col = "red", size = 10))
rgl::planes3d(cent[1],cent[2],cent[3],-sum(cent^2), alpha = 0.5, col = "red")
finalize3D()

ini3D()
cent <- c(100,100,100)
r <- 75
planeC <- c(cent+r/3)
planeC <- c(planeC, -sum(planeC^2))
p <- genNDSet(1000, keep = TRUE,
  argsSphere = list(center = cent, radius = r, below = FALSE, plane = planeC))
rgl::spheres3d(cent, radius=r, color = "grey100", alpha=0.1)
plotPoints3D(p)
plotPoints3D(p[!p$dom,], argsPlot3d = list(col = "red", size = 10))
rgl::planes3d(planeC[1],planeC[2],planeC[3],planeC[4], alpha = 0.5, col = "red")
finalize3D()
}
