% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_doc.R
\docType{package}
\name{gMOIP}
\alias{gMOIP}
\alias{gMOIP-package}
\title{2D plots of linear/integer programming models.}
\description{
Make 2D and 3D plots of the polytope of a linear programming (LP), integer linear programming
(ILP) model, or mixed integer linear programming (MILP) model with 2 or 3 variables, including
integer points, ranges and iso profit curve.
}
\details{
Can also make a plot of the bi-objective criterion space and the non-dominated (Pareto) set for
bi-objective LP/ILP/MILP programming models.
}
\seealso{
\link{plotPolytope} and \link{plotCriterion2D}.
}
\author{
Lars Relund <lars@relund.dk> [aut, cre].
}
\keyword{package}
