% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{gMOIPTheme}
\alias{gMOIPTheme}
\title{ggPlot theme for the package}
\usage{
gMOIPTheme(...)
}
\arguments{
\item{...}{Further arguments parsed to \link[ggplot2:theme]{ggplot2::theme}.}
}
\value{
The theme object.
}
\description{
ggPlot theme for the package
}
