% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotCones3D}
\alias{plotCones3D}
\title{Plot a cone defined by a point in 3D.}
\usage{
plotCones3D(
  pts,
  drawPoint = TRUE,
  drawLines = TRUE,
  reverse = FALSE,
  rectangle = FALSE,
  ...
)
}
\arguments{
\item{pts}{A matrix with a point in each row.}

\item{drawPoint}{Draw the points defining the cone.}

\item{drawLines}{Draw lines of the cone.}

\item{reverse}{Cones are defined as the point minus R3+}

\item{rectangle}{Draw the cone as a rectangle.}

\item{...}{Further arguments passed on the the rgl plotting functions. This must be done as
lists (see examples). Currently the following arguments are supported:
\itemize{
\item \code{argsPlot3d}: A list of arguments for \code{\link[rgl:plot3d]{rgl::plot3d}}.
\item \code{argsSegments3d}: A list of arguments for \code{\link[rgl:points3d]{rgl::segments3d}}.
\item \code{argsPolygon3d}: A list of arguments for \code{\link[rgl:polygon3d]{rgl::polygon3d}}.
}}
}
\value{
NULL (invisible)
}
\description{
The cones are defined as the point plus R3+.
}
\examples{
ini3D(argsPlot3d = list(xlim = c(0,6), ylim = c(0,6), zlim = c(0,6)))
plotCones3D(c(4,4,4), drawLines = FALSE, drawPoint = TRUE,
           argsPlot3d = list(col = "red", size = 10),
           argsPolygon3d = list(alpha = 1), rectangle = TRUE)
plotCones3D(c(1,1,1), reverse = TRUE, rectangle = TRUE)
plotCones3D(matrix(c(3,3,3,2,2,2), ncol = 3, byrow = TRUE))
finalize3D()
}
